fix conflict image compress
